mod STEP8

+++ pico_v2_STEP8_micromouse/run.h
@@ -61,7 +61,7 @@ public:
   void straight(int len, int init_speed, int max_sp, int finish_speed);
   void accelerate(int len, int finish_speed);
   void oneStep(int len, int init_speed);
-  void decelerate(int len, float init_speed);
+  void decelerate(int len, int init_speed);
   void rotate(t_local_direction dir, int times);
 
 private:
